Type Species shared slice state
